dradis: Only access the memory actually used
So far, we were accessing the entire buffer content, even if the driver
filled only a portion of it.

We should only access what v4l2_buffer::bytesperused describes. We still
assume that we don't have a stride larger than lenght * depth, but I
guess we can figure it out later if we ever encounter that case.

Signed-off-by: Maxime Ripard <mripard@kernel.org>